Add Modals for Auth/Fix Bug that fills form/README
I added modals for the auth selections as well as the view games
selection.  There was a bug that when O would win sometimes, it would
fill every cell with an O upon clicking play new game.  I needed to add
a .html line to both conditions on onPlayGame in events.  Fixed some
.show and .hide options.  Updated the README and updated styling.